refactoring of integration tests of basic example
